pyproject.toml
[project]
name = "ai-news2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "aiohttp>=3.12.9",
    "alembic>=1.16.1",
    "arxiv>=2.2.0",
    "google-genai>=1.19.0",
    "gtts>=2.5.4",
    "psutil>=7.0.0",
    "pydantic>=2.11.5",
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=6.1.1",
    "pyyaml>=6.0.2",
    "rich>=14.0.0",
    "sqlalchemy>=2.0.41",
    "tenacity>=9.1.2",
]
